deleted files for test, .idea folder, created Adapter.java
