+++ app/build.gradle
@@ -22,13 +22,6 @@ dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation 'com.android.support:appcompat-v7:28.0.0'
     implementation 'com.android.support.constraint:constraint-layout:1.1.3'
-    testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support.test:runner:1.0.2'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
-
-    //Retrofit
     implementation 'com.squareup.retrofit2:retrofit:2.4.0'
-
-    //Retrofit JSON
     implementation 'com.squareup.retrofit2:converter-gson:2.4.0'
 }
