+++ app/src/main/java/ua/lpnu/yana/apilab/MainActivity.java
@@ -2,58 +2,14 @@ package ua.lpnu.yana.apilab;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.util.Log;
-import android.widget.Toast;
-
-import java.util.ArrayList;
-
-import retrofit2.Call;
-import retrofit2.Callback;
-import retrofit2.Response;
-import retrofit2.Retrofit;
-import retrofit2.converter.gson.GsonConverterFactory;
 
 public class MainActivity extends AppCompatActivity {
-
-    private static final String TAG = "MainActivity";
-    private static final String BASE_URL = "https://api.pokemontcg.io/v1/";
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-
-        Retrofit retrofit = new Retrofit.Builder()
-                .baseUrl(BASE_URL)
-                .addConverterFactory(GsonConverterFactory.create())
-                .build();
-
-        PokemonAPI pokemonAPI = retrofit.create(PokemonAPI.class);
-        Call<Feed> call = pokemonAPI.getData();
-
-        call.enqueue(new Callback<Feed>() {
-
-            @Override
-            public void onResponse(Call<Feed> call, Response<Feed> response) {
-                Log.d(TAG, "onResponse: Server Response: " + response.toString());
-                Log.d(TAG, "onResponse: received information: " + response.body().toString());
-
-                ArrayList<Cards> cardsList = response.body().getCards();
-                for (Cards card : cardsList) {
-
-                    Log.d(TAG, "id: " + card.getId()+ "\n" +
-                                    "name: " + card.getName()+ "\n" +
-                                    "image: " + card.getImageUrlHiRes()+ "\n" +
-                                    "number: " + card.getNationalPokedexNumber()+ "\n" +
-                                    "_____________________________\n\n");
-                }
-            }
-
-            @Override
-            public void onFailure(Call<Feed> call, Throwable t) {
-                Log.e(TAG, "onFailure: Something went wrong: " + t.getMessage());
-                Toast.makeText(MainActivity.this, "Something went wrong", Toast.LENGTH_SHORT).show();
-            }
-        });
+        Adapter adapter = new Adapter();
+        adapter.makeRequest();
     }
 }
